+++ features/support/helpers/transaction_id_generator.rb
@@ -2,7 +2,7 @@ module KnowsAboutTransactionIdGeneration
   attr_accessor :last_request_id
 
   def new_request_id
-    @last_request_id = 'test-id-' + SecureRandom.uuid
+    @last_request_id = SecureRandom.uuid
   end
 end
 World(KnowsAboutTransactionIdGeneration)